+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.5 [2019-02-19] 
+
+* Tested on Keycloak 4.8.3
+* Long-Lived tokens are restricted depending on a realm role
+
 ## 0.4 [2019-02-13] 
 
 * Support for long-lived tokens